requirements.txt
loguru==0.6.0
frozenlist==1.3.3
alembic==1.9.1
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
click==8.1.3
charset-normalizer==3.0.1
multidict==6.0.4
requests==2.28.1
requests-toolbelt==0.10.1
yarl==1.8.2
greenlet==2.0.1
idna==3.4
attrs==22.2.0
chardet==5.1.0
asgiref==3.6.0
dominate==2.7.0
pypiwin32==223
colorama==0.4.6
win32-setctime==1.1.0
jwt==1.3.1
itsdangerous==2.1.2
brotli==1.0.9
Jinja2==3.1.2
Werkzeug==2.2.2
Flask[async]==2.2.2
Flask-wtf==1.0.1
Flask-restful==0.3.9
Flask-SQLAlchemy==3.0.2
Flask-Caching==2.0.1
Flask-Migrate==4.0.0
Flask-Alembic==2.0.1
sqlalchemy==1.4.45
sqlalchemy-migrate==0.13.0
pycryptodomex==3.16.0
pycryptodome==3.16.0
StrEnum==0.4.9
python-dateutil==2.8.2